added responsiv navbar

+++ src/pages/Dashboard/Admin/ManageProduct/ManageProduct.js
@@ -6,19 +6,37 @@ import MuiButton from '../../../../StyledComponent/MuiButton';
 
 const ManageProduct = ({ product }) => {
   const { name, price, _id } = product;
-  // const [deleteId, setDeleteId] = useState();
+  const Swal = require('sweetalert2')
 
   const handleClick = (id) => {
-    // setDeleteId(id);
-    fetch(`http://localhost:5000/deleteProducts/${id}`, {
-      method: 'DELETE',
-      headers: { 'Content-Type': 'application/json' }
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`http://localhost:5000/deleteProducts/${id}`, {
+          method: 'DELETE',
+          headers: { 'Content-Type': 'application/json' }
+        })
+          .then(res => res.json())
+          .then(result => {
+            if (result.acknowledged) {
+              Swal.fire(
+                'Deleted!',
+                'Your file has been deleted.',
+                'success'
+              )
+            }
+          })
+      }
     })
-      .then(res => res.json())
-      .then(result => {
-        console.log(result)
-      })
   }
+  
   return (
     <TableBody>
       <TableRow>
